% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic statistics for population and loci.R
\docType{data}
\name{BS}
\alias{BS}
\title{Displaying per locus information}
\format{
An object of class \code{basic.stats} of length 7.
}
\usage{
BS
}
\description{
Summarizes the locus information with its genetic diversity details
}
\keyword{datasets}
