% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic statistics for population and loci.R
\docType{data}
\name{gac}
\alias{gac}
\title{Generating multi locus genotype accumulation curve}
\format{
An object of class \code{matrix} with 1000 rows and 17 columns.
}
\usage{
gac
}
\description{
Checks how much power the used loci have to identify unique MLGs
}
\keyword{datasets}
