% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic statistics for population and loci.R
\docType{data}
\name{l.col}
\alias{l.col}
\title{Displaying DAPC to categorize the populations in different clusters}
\format{
An object of class \code{character} of length 40.
}
\usage{
l.col
}
\description{
Gives information about how the populations are related and clustered across two principal components (PCs)
}
\keyword{datasets}
