% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic statistics for population and loci.R
\docType{data}
\name{tks}
\alias{tks}
\title{Loading Genalex datafile}
\format{
An object of class \code{genclone} of length 1.

An object of class \code{genclone} of length 1.
}
\usage{
tks

tks
}
\description{
Generates file with details about individuals, loci and allele sizes

Summarizes if there exists some genetic structure in populations or not
}
\keyword{datasets}
