% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic statistics for population and loci.R
\docType{data}
\name{tks2}
\alias{tks2}
\title{Displaying Index of Association}
\format{
An object of class \code{genind} of length 1.
}
\usage{
tks2
}
\description{
Checks how the populations are associated using IoA
}
\keyword{datasets}
