% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic statistics for population and loci.R
\docType{data}
\name{tks_sub}
\alias{tks_sub}
\title{Displaying minimum spanning network}
\format{
An object of class \code{genind} of length 1.
}
\usage{
tks_sub
}
\description{
Summarizes how the populations used are spanned in a network
}
\keyword{datasets}
