% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic statistics for population and loci.R
\docType{data}
\name{tkspair}
\alias{tkspair}
\title{Displaying Linkage Disequillibrium over all loci}
\format{
An object of class \code{pairia} (inherits from \code{matrix}) with 153 rows and 2 columns.
}
\usage{
tkspair
}
\description{
Summarizes if the used loci are linked or not
}
\keyword{datasets}
